Moved NRML to hazardlib
