+++ openquake/calculators/tests/classical_test.py
@@ -18,7 +18,7 @@
 
 from nose.plugins.attrib import attr
 from openquake.baselib import parallel
-from openquake.commonlib import InvalidFile
+from openquake.hazardlib import InvalidFile
 from openquake.calculators.export import export
 from openquake.calculators.tests import CalculatorTestCase, check_platform
 from openquake.qa_tests_data.classical import (
