+++ openquake/commands/reduce.py
@@ -20,10 +20,9 @@
 from __future__ import print_function
 import random
 import shutil
-from openquake.risklib import valid
+from openquake.hazardlib import valid, nrml
 from openquake.baselib.python3compat import encode
 from openquake.baselib import sap
-from openquake.commonlib import nrml
 
 
 def random_filter(objects, reduction_factor, seed=42):
