+++ openquake/commonlib/hazard_writers.py
@@ -30,7 +30,7 @@ import numpy
 from xml.etree import ElementTree as et
 
 from openquake.baselib.node import Node, scientificformat, floatformat
-from openquake.commonlib import nrml
+from openquake.hazardlib import nrml
 
 by_imt = operator.itemgetter('imt', 'sa_period', 'sa_damping')
