+++ openquake/commonlib/risk_writers.py
@@ -26,10 +26,10 @@ import numpy
 
 from xml.etree import ElementTree as et
 
-from openquake.commonlib.nrml import SERIALIZE_NS_MAP
+from openquake.hazardlib.nrml import SERIALIZE_NS_MAP
 from openquake.baselib.general import groupby, writetmp
 from openquake.baselib.node import Node
-from openquake.commonlib import nrml
+from openquake.hazardlib import nrml
 from openquake.commonlib.writers import FIVEDIGITS
