+++ openquake/commonlib/shapefileparser.py
@@ -28,9 +28,9 @@ from collections import OrderedDict
 from openquake.baselib.general import groupby
 from openquake.hazardlib import geo
 from openquake.hazardlib.geo.surface import SimpleFaultSurface
-from openquake.commonlib import nrml
+from openquake.hazardlib import nrml
 from openquake.baselib.node import Node, striptag
-from openquake.commonlib.sourceconverter import SourceConverter
+from openquake.hazardlib.sourceconverter import SourceConverter
 
 # maximum field size allowed by shapefile
 FIELD_SIZE = 255
