+++ openquake/commonlib/tests/_utils.py
@@ -20,7 +20,7 @@ from xml.etree.ElementTree import parse
 
 from openquake.baselib.general import writetmp
 from openquake.baselib.node import tostring
-from openquake.commonlib.nrml import PARSE_NS_MAP
+from openquake.hazardlib.nrml import PARSE_NS_MAP
 
 
 def get_path(fname_or_fileobject):
