+++ openquake/commonlib/tests/calc_test.py
@@ -3,8 +3,10 @@ import tempfile
 import unittest
 import numpy
 from openquake.baselib import hdf5, general
-from openquake.commonlib import nrml_examples, calc, source, nrml
-from openquake.commonlib.sourceconverter import SourceConverter, RuptureConverter
+from openquake.hazardlib import nrml
+from openquake.hazardlib.sourceconverter import (
+    SourceConverter, RuptureConverter)
+from openquake.commonlib import nrml_examples, calc, source
 
 NRML_DIR = os.path.dirname(nrml_examples.__file__)
 MIXED_SRC_MODEL = os.path.join(NRML_DIR, 'source_model/mixed.xml')
