+++ openquake/commonlib/tests/nrml_test.py
@@ -19,8 +19,8 @@
 import unittest
 import io
 from openquake.baselib.general import writetmp
-from openquake.commonlib.sourceconverter import SourceConverter
-from openquake.commonlib.nrml import read, parse, node_to_xml, get_tag_version
+from openquake.hazardlib.sourceconverter import SourceConverter
+from openquake.hazardlib.nrml import read, parse, node_to_xml, get_tag_version
 
 
 class NrmlTestCase(unittest.TestCase):
