+++ openquake/commonlib/tests/riskmodels_test.py
@@ -23,7 +23,8 @@ import mock
 import numpy
 from numpy.testing import assert_almost_equal
 from openquake.baselib.general import writetmp
-from openquake.commonlib import InvalidFile, nrml, nrml_examples, riskmodels
+from openquake.hazardlib import InvalidFile, nrml
+from openquake.commonlib import nrml_examples, riskmodels
 from openquake.qa_tests_data.scenario_damage import case_4b
 
 EXAMPLES_DIR = os.path.dirname(nrml_examples.__file__)
