+++ openquake/engine/tools/correct_complex_sources.py
@@ -22,7 +22,7 @@ Correct complex fault sources to comply with Aki and Richards convention.
 import sys
 import numpy
 
-from openquake.commonlib import nrml
+from openquake.hazardlib import nrml
 from openquake.baselib.node import node_from_xml
 
 from openquake.hazardlib.geo.line import Line
